DD-121-fix: 지원자목록 상태변경 후 바로 업데이트 안되는 에러 수정
